issue:add question, meeting, category

+++ grpc/base.go
@@ -49,20 +49,20 @@ func parseString(src string, sep string) (string, int) {
 	if len(arr) < 2 {
 		return "", -1
 	}
-	st,er := strconv.ParseInt(arr[1], 10, 32)
+	st, er := strconv.ParseInt(arr[1], 10, 32)
 	if er != nil {
 		return "", -1
 	}
 	return arr[0], int(st)
 }
 
-func parseStringToInt(src string) int {
+func parseStringToInt(src string) int64 {
 	if src == "" {
 		return -1
 	}
-	st,er := strconv.ParseInt(src, 10, 32)
+	st, er := strconv.ParseInt(src, 10, 32)
 	if er != nil {
 		return -1
 	}
-	return int(st)
+	return st
 }
